form_app/pyproject.toml
[project]
name = "form-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "gunicorn>=23.0.0",
    "itsdangerous>=2.2.0",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.2.9",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "shared",
]

[dependency-groups]
dev = [
    "debugpy>=1.8.19",
]

[tool.uv.sources]
shared = { workspace = true }
